requirements.txt
crewai==0.105.0
crewai_tools==0.37.0
litellm==1.60.2
openai==1.66.3
python-dotenv==1.0.1
streamlit==1.25.0
Pillow==9.4.0
pysqlite3-binary
